test(maestro): split inlay hint tests into focused submodules
