+++ crates/vize_maestro/src/ide/inlay_hint.rs
@@ -10,6 +10,13 @@ mod expr_regions;
 mod script;
 mod template;
 
+#[cfg(test)]
+mod binding_hint_tests;
+#[cfg(test)]
+mod i18n_hint_tests;
+#[cfg(test)]
+mod prop_hint_tests;
+
 use tower_lsp::lsp_types::{InlayHint, Position, Range, Url};
 use vize_croquis::{Drawer, DrawerOptions};
 
@@ -228,440 +235,5 @@ impl InlayHintService {
 
     fn is_ident_char(c: u8) -> bool {
         c.is_ascii_alphanumeric() || c == b'_' || c == b'$'
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::fs;
-
-    use super::InlayHintService;
-    use tower_lsp::lsp_types::{InlayHintLabel, Position, Range, Url};
-
-    #[test]
-    fn test_props_destructure_analysis() {
-        let content = r#"<script setup lang="ts">
-const { title, disabled } = defineProps<{
-  title: string
-  disabled?: boolean
-}>()
-
-console.log(title)
-</script>
-
-<template>
-  <div>{{ title }}</div>
-</template>"#;
-
-        let uri = Url::parse("file:///test.vue").unwrap();
-        let range = Range {
-            start: Position {
-                line: 0,
-                character: 0,
-            },
-            end: Position {
-                line: 100,
-                character: 0,
-            },
-        };
-
-        let hints = InlayHintService::get_hints(content, &uri, range);
-
-        // Should have hints for title in script (line 6) and template (line 10)
-        assert!(!hints.is_empty(), "Should have inlay hints");
-
-        // Verify all hints are #props.
-        for hint in &hints {
-            if let InlayHintLabel::String(label) = &hint.label {
-                assert_eq!(label, "#props.");
-            }
-        }
-    }
-
-    #[test]
-    fn template_literal_text_never_hints_prop_names() {
-        let content = r#"<script setup lang="ts">
-const { size, state, tone, variant } = defineProps<{
-  size: string
-  state: string
-  tone: string
-  variant: string
-}>()
-</script>
-
-<template>
-  <span
-    :aria-disabled="state === 'disabled'"
-    :class="[
-      'tag',
-      `tag--size-${size}`,
-      `tag--state-${state}`,
-      `tag--tone-${tone}`,
-      `tag--variant-${variant}`,
-    ]"
-  >
-    <slot />
-  </span>
-</template>"#;
-
-        let uri = Url::parse("file:///test.vue").unwrap();
-        let range = Range {
-            start: Position {
-                line: 0,
-                character: 0,
-            },
-            end: Position {
-                line: 100,
-                character: 0,
-            },
-        };
-
-        let hints = InlayHintService::get_hints(content, &uri, range);
-
-        // Exactly one hint per real reference: the aria-disabled comparison
-        // plus the four `${...}` interpolations. The static `tag--size-`
-        // template text must not hint.
-        assert_eq!(hints.len(), 5, "{hints:#?}");
-
-        let lines: Vec<&str> = content.lines().collect();
-        for hint in &hints {
-            let line = lines[hint.position.line as usize];
-            let before = &line[..hint.position.character as usize];
-            assert!(
-                before.ends_with("${") || before.ends_with("=\"") || before.ends_with("\""),
-                "hint must sit on a code reference, not literal text: line {:?} col {} ({before:?})",
-                hint.position.line,
-                hint.position.character
-            );
-        }
-    }
-
-    #[test]
-    fn test_props_destructure_with_alias() {
-        let content = r#"<script setup lang="ts">
-const { title: localTitle } = defineProps<{
-  title: string
-}>()
-
-console.log(localTitle)
-</script>
-
-<template>
-  <div>{{ localTitle }}</div>
-</template>"#;
-
-        let uri = Url::parse("file:///test.vue").unwrap();
-        let range = Range {
-            start: Position {
-                line: 0,
-                character: 0,
-            },
-            end: Position {
-                line: 100,
-                character: 0,
-            },
-        };
-
-        let hints = InlayHintService::get_hints(content, &uri, range);
-
-        // Should have hints for localTitle (the alias), not title
-        assert!(
-            !hints.is_empty(),
-            "Should have inlay hints for aliased prop"
-        );
-    }
-
-    #[test]
-    fn test_no_hints_in_define_props_type() {
-        let content = r#"<script setup lang="ts">
-const { title } = defineProps<{
-  title: string
-}>()
-</script>
-
-<template>
-  <div>{{ title }}</div>
-</template>"#;
-
-        let uri = Url::parse("file:///test.vue").unwrap();
-        let range = Range {
-            start: Position {
-                line: 0,
-                character: 0,
-            },
-            end: Position {
-                line: 100,
-                character: 0,
-            },
-        };
-
-        let hints = InlayHintService::get_hints(content, &uri, range);
-
-        // Check that no hints are in the defineProps type definition
-        // (lines 1-3 in script, which is around line 1-4 in the file)
-        for hint in &hints {
-            assert!(
-                hint.position.line > 3,
-                "Hint should not be in defineProps type definition, found at line {}",
-                hint.position.line
-            );
-        }
-    }
-
-    #[test]
-    fn test_is_in_string() {
-        assert!(!InlayHintService::is_in_string("foo bar", 4));
-        assert!(InlayHintService::is_in_string("'foo bar'", 4));
-        assert!(InlayHintService::is_in_string("\"foo bar\"", 4));
-        assert!(!InlayHintService::is_in_string("\"foo\" bar", 6));
-        assert!(InlayHintService::is_in_string("`foo bar`", 4));
-    }
-
-    #[test]
-    fn test_reactive_binding_inlay_hint() {
-        let content = r#"<script setup lang="ts">
-import { ref, computed } from 'vue'
-const count = ref(0)
-const doubled = computed(() => count.value * 2)
-</script>
-"#;
-        let uri = Url::parse("file:///reactive.vue").unwrap();
-        let range = Range {
-            start: Position {
-                line: 0,
-                character: 0,
-            },
-            end: Position {
-                line: 100,
-                character: 0,
-            },
-        };
-        let hints = InlayHintService::get_hints(content, &uri, range);
-        let labels: Vec<String> = hints
-            .iter()
-            .filter_map(|h| match &h.label {
-                InlayHintLabel::String(s) => Some(s.clone()),
-                _ => None,
-            })
-            .collect();
-        assert!(
-            labels.iter().any(|s| s.contains("Ref")),
-            "expected a Ref<...> inlay hint, got {labels:?}",
-        );
-        assert!(
-            labels.iter().any(|s| s.contains("ComputedRef")),
-            "expected a ComputedRef<...> inlay hint, got {labels:?}",
-        );
-    }
-
-    #[test]
-    fn test_reactive_binding_inlay_hint_resolves_inner_type() {
-        // Follow-up to #696: the inlay hint must surface the inferred
-        // value type rather than a placeholder. `ref(0)` is number;
-        // `ref<string>()` carries an explicit type parameter.
-        let content = r#"<script setup lang="ts">
-import { ref } from 'vue'
-const counter = ref(0)
-const label = ref<string>()
-</script>
-"#;
-        let uri = Url::parse("file:///inferred.vue").unwrap();
-        let range = Range {
-            start: Position {
-                line: 0,
-                character: 0,
-            },
-            end: Position {
-                line: 100,
-                character: 0,
-            },
-        };
-        let hints = InlayHintService::get_hints(content, &uri, range);
-        let labels: Vec<String> = hints
-            .iter()
-            .filter_map(|h| match &h.label {
-                InlayHintLabel::String(s) => Some(s.clone()),
-                _ => None,
-            })
-            .collect();
-        assert!(
-            labels.iter().any(|s| s.contains("Ref<number>")),
-            "expected Ref<number> inlay hint for ref(0), got {labels:?}",
-        );
-        assert!(
-            labels.iter().any(|s| s.contains("Ref<string>")),
-            "expected Ref<string> inlay hint for ref<string>(), got {labels:?}",
-        );
-    }
-
-    #[test]
-    fn test_no_hints_in_event_name_pattern() {
-        // Test that "title" in "update:title" event name does not get a hint
-        let content = r#"<script setup lang="ts">
-const { title } = defineProps<{
-  title: string
-}>()
-
-const emit = defineEmits<{
-  (e: 'update:title', value: string): void
-}>()
-</script>
-
-<template>
-  <input :value="title" @update:title="emit('update:title', $event)" />
-</template>"#;
-
-        let uri = Url::parse("file:///test.vue").unwrap();
-        let range = Range {
-            start: Position {
-                line: 0,
-                character: 0,
-            },
-            end: Position {
-                line: 100,
-                character: 0,
-            },
-        };
-
-        let hints = InlayHintService::get_hints(content, &uri, range);
-
-        // Should have hints for title in :value="title" and possibly template
-        // But NOT for title in 'update:title' event names
-        for hint in &hints {
-            // Get the position in the content
-            let line = hint.position.line as usize;
-            let lines: Vec<&str> = content.lines().collect();
-            if line < lines.len() {
-                let line_content = lines[line];
-                // Verify the hint is not on a line containing 'update:title' pattern
-                // where title immediately follows a colon
-                let char_pos = hint.position.character as usize;
-                if char_pos > 0 && char_pos <= line_content.len() {
-                    let before_char = line_content.as_bytes().get(char_pos - 1);
-                    assert_ne!(
-                        before_char,
-                        Some(&b':'),
-                        "Hint should not be placed after colon (event name pattern)"
-                    );
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_props_without_destructure_in_template() {
-        // Test that props defined without destructuring also get hints in template
-        let content = r#"<script setup lang="ts">
-const props = defineProps<{
-  title: string
-  count: number
-}>()
-
-// In script, we access via props.title (no hint needed for 'title' alone)
-console.log(props.title)
-</script>
-
-<template>
-  <div>{{ title }}</div>
-  <span>{{ count }}</span>
-</template>"#;
-
-        let uri = Url::parse("file:///test.vue").unwrap();
-        let range = Range {
-            start: Position {
-                line: 0,
-                character: 0,
-            },
-            end: Position {
-                line: 100,
-                character: 0,
-            },
-        };
-
-        let hints = InlayHintService::get_hints(content, &uri, range);
-
-        // Should have hints for title and count in template (lines 11 and 12)
-        // Even though props are not destructured
-        let template_hints: Vec<_> = hints.iter().filter(|h| h.position.line >= 11).collect();
-
-        assert!(
-            !template_hints.is_empty(),
-            "Should have hints for props in template even without destructuring"
-        );
-    }
-
-    #[test]
-    fn test_i18n_message_preview_without_script_setup() {
-        let content = r#"<template>
-  <p>{{ $t("auth.login") }}</p>
-</template>
-<i18n lang="json">
-{
-  "en": { "auth": { "login": "Log in" } }
-}
-</i18n>
-"#;
-
-        let uri = Url::parse("file:///test.vue").unwrap();
-        let range = Range {
-            start: Position {
-                line: 0,
-                character: 0,
-            },
-            end: Position {
-                line: 100,
-                character: 0,
-            },
-        };
-
-        let hints = InlayHintService::get_hints(content, &uri, range);
-        let labels: Vec<&str> = hints
-            .iter()
-            .filter_map(|hint| match &hint.label {
-                InlayHintLabel::String(label) => Some(label.as_str()),
-                _ => None,
-            })
-            .collect();
-
-        assert_eq!(labels, vec!["= Log in"]);
-    }
-
-    #[test]
-    fn test_i18n_message_preview_from_workspace_json_catalog() {
-        let dir = tempfile::tempdir().unwrap();
-        let source_path = dir.path().join("src/components/LoginButton.vue");
-        let locale_path = dir.path().join("src/locales/en.json");
-        fs::create_dir_all(source_path.parent().unwrap()).unwrap();
-        fs::create_dir_all(locale_path.parent().unwrap()).unwrap();
-        fs::write(&locale_path, r#"{ "auth": { "login": "Log in" } }"#).unwrap();
-
-        let content = r#"<template>
-  <p>{{ $t("auth.login") }}</p>
-</template>
-"#;
-        fs::write(&source_path, content).unwrap();
-
-        let uri = Url::from_file_path(&source_path).unwrap();
-        let range = Range {
-            start: Position {
-                line: 0,
-                character: 0,
-            },
-            end: Position {
-                line: 100,
-                character: 0,
-            },
-        };
-
-        let hints = InlayHintService::get_hints(content, &uri, range);
-        let labels: Vec<&str> = hints
-            .iter()
-            .filter_map(|hint| match &hint.label {
-                InlayHintLabel::String(label) => Some(label.as_str()),
-                _ => None,
-            })
-            .collect();
-
-        assert_eq!(labels, vec!["= Log in"]);
     }
 }
